Chebyshev Interpolation

Chebyshev interpolation is an extremely accurate method for interpolating smooth functions. However, the extremely high accuracy is achieved by careful placement of the interpolation nodes at the Chebyshev nodes. Chebyshev interpolation is a good way to approximate a smooth function that is very expensive to evaluate. By evaluating the function at the Chebyshev nodes, which are given by the chebygrid function, highly accurate and efficient interpolators can be generated to speed things up. This is the only part of BasicInterpolators that is not really so basic, but chebyshev interpolation is a powerful tool in the right context and it's worth providing easy access to it.

To create a ChebyshevInterpolator or BichebyshevInterpolator, it is probably easiest to pass a function to the constructor, which will evaluate the function at the Chebyshev nodes and generate interpolating coefficients. For example, for two-dimensional interpolation of $f(x,y) = \sin(x) + \cos(y)$ with $x,y ∈ [-3,3]$

julia> using BasicInterpolators: BichebyshevInterpolator

julia> f(x,y) = sin(x) + cos(y)
f (generic function with 1 method)

julia> P = BichebyshevInterpolator(f, -3, 3, 16, -3, 3, 16);

julia> P(1, 2) - f(1, 2)
1.7115198147621413e-12

With 16 points in each dimension, the error is already about one part per trillion. This is a contrived example and the incredible accuracy of Chebyshev interpolation can be demonstrated more thoroughly, but the important fact is that Chebyshev approximations to smooth functions have exponential, or "infinite order" convergence. As soon as there are enough points for the Chebyshev approximation to capture the function's basic shape, the error will plummet to machine precision as more points are added. For more on this and related topics, I recommend the following book

Boyd, John P. Chebyshev and Fourier spectral methods. Courier Corporation, 2001.

Warning

The Chebyshev interpolators are not thread-safe. They compute cosine expansions and, in the 2D case, do some linear algebra in-place using arrays stored with the objects. A single chebyshev interpolator should never be called by multiple threads at once.

As the warning above indicates, the Chebyshev interpolators do some things in the background that the other interpolators don't. I put some effort into making them fast (at least once constructed) by avoiding as many direct cos evaluations as possible and doing the unavoidable calculations in-place using fast linear algebra functions (mul!). I think the speed of the Chebyshev interpolators here should compare well with any other implementations, but they should be considerably easier to use.

BasicInterpolators.chebygrid — Function

chebygrid(n::Int)

Create an array of n chebyshev nodes in [-1,1]

chebygrid(xa::Real, xb::Real, n::Int)

Create an array of n chebyshev nodes in [xa,xb]

chebygrid(xa::Real, xb::Real, nx::Int, ya::Real, yb::Real, ny::Int

Create a two-dimensional grid of chebyshev nodes using nx points along the first axis, in [xa,xb], and ny points along the second axis, in [ya,yb].

BasicInterpolators.ChebyshevInterpolator — Type

ChebyshevInterpolator(x::AbstractVector{<:Real}, y::AbstractVector{<:Real})

Construct a ChebyshevInterpolator for the points in x and y. The x coordinates must be arranged on a chebyshev grid, which can be generated using the chebygrid function.

ChebyshevInterpolator(f::Function, xa::Real, xb::Real, n::Int)

Construct a ChebyshevInterpolator for the function f using n function evaluations in the range [xa,xb]. The function evaluations will occur on the chebyshev nodes.

BasicInterpolators.BichebyshevInterpolator — Type

BichebyshevInterpolator(x::AbstractVector{<:Real},
                        y::AbstractVector{<:Real},
                        Z::AbstractArray{<:Real,2})

Construct a BichebyshevInterpolator for the grid of points with axis 1 coordinates in x, axis 2 coordinates in y, and grid values in Z. The given points must lie on a chebyshev grid in each direction. These can be generated with the chebygrid function or the interpolator can be constructed directly from a function using the method below.

BichebyshevInterpolator(f::Function,
                        xa::Real, xb::Real, nx::Int,
                        ya::Real, yb::Real, ny::Int)

Construct a BichebyshevInterpolator for the function f using a grid of nx points on the first axis in [xa,xb] and ny points on the second axis in [ya,yb].