Cubic Interpolation

The types below implement piecewise cubic interpolation in one and two dimensions. These are not typical cubic splines, which are also implemented in this package. The CubicInterpolator and BicubicInterpolator do not guarantee smoothness at the interpolation nodes. However, they can be used with unevenly spaced grids and perform well when interpolating smooth curves/surfaces. For cubic splines, look here.

BasicInterpolators.CubicInterpolator — Type

CubicInterpolator(x::AbstractVector{<:Real}, y::AbstractVector{<:Real})

Construct a CubicInterpolator for the points in x and y

CubicInterpolator(f::Function, xa::Real, xb::Real, n::Int)

Construct a CubicInterpolator for the function f using n evenly spaced function evaluations in the range [xa,xb]

BasicInterpolators.BicubicInterpolator — Type

BicubicInterpolator(x::AbstractVector{<:Real},
                    y::AbstractVector{<:Real},
                    Z::AbstractArray{<:Real,2})

Construct a BicubicInterpolator for the grid of points with axis 1 coordinates in x, axis 2 coordinates in y, and grid values in Z.

BicubicInterpolator(f::Function,
                    xa::Real, xb::Real, nx::Int,
                    ya::Real, yb::Real, ny::Int)

Construct a BicubicInterpolator for the function f using a grid of nx points evenly spaced on the first axis in [xa,xb] and ny points evenly spaced on the second axis in [ya,yb].