Miscellaneous

Pieces of the package that might be useful 🤷

BasicInterpolators.findcell — Function

findcell(q, V)

Use bisection search to find the cell containing q, assuming V is a sorted vector of coordinates. The returned integer is the index of the element in V immediately less than q. For example, if findcell returns 2, then q ∈ [V[2],V[3]). If q is less than every element in V, 1 is returned, indicating the first cell in V. If q is greater than every element in V, length(V)-1 is returned, indicating the last cell in V.

BasicInterpolators.chebygrid — Function

chebygrid(n)

Create an array of n chebyshev nodes in [-1,1]

chebygrid(xa, xb, n)

Create an array of n chebyshev nodes in [xa,xb]

chebygrid(xa, xb, nx, ya, yb, ny)

Create a two-dimensional grid of chebyshev nodes using nx points along the first axis, in [xa,xb], and ny points along the second axis, in [ya,yb].