General Polynomial Interpolation in One Dimension

For general polynomial interpolation in one dimension, BasicInterpolators provides a few implementations of Neville's Algorithm.

For quadratic interpolation of any three points, use the quadratic function.

For cubic interpolation of any four points, use the cubic function.

For nth order polynomial interpolation of arbitrary points, the neville function is available.

All three functions use the same basic algorithm, but the quadratic and cubic versions are written out explicitly for those cases and will be faster than the general neville function.

BasicInterpolators.quadratic — Function

quadratic(x::Real, xp::AbstractVector{<:Real}, yp::AbstractVector{<:Real})

Perform quadratic polynomial interpolation of xp and yp at the point x, using Neville's algorithm. xp and yp must both contain three points.

BasicInterpolators.cubic — Function

cubic(x::Real, xp::AbstractVector{<:Real}, yp::AbstractVector{<:Real})

Perform cubic polynomial interpolation of xp and yp at the point x, using Neville's algorithm. xp and yp must both contain four points.

BasicInterpolators.neville — Function

neville(x::Real, xp::AbstractVector{<:Real}, yp::AbstractVector{<:Real})

Perform polynomial interpolation of xp and yp at the point x, using Neville's algorithm with as many points as are provided. xp and yp must have the same length. With only 3 or 4 points (quadratic or cubic polynomial interpolation) the quadratic and cubic functions will be faster.