Cubic Spline Interpolation

The types below implement piecewise cubic spline interpolation in one and two dimensions. Both types guarantee smoothness at the interpolation nodes. The CubicSplineInterpolator is a natural spline, with the second derivative set to zero at the boundaries, and it can be used with unevenly spaced points. The BicubicSplineInterpolator requires points to be evenly spaced in both dimensions and patches together cubic polynomials using finite-difference approximations of the surface's derivatives at the interpolation nodes.

BasicInterpolators.CubicSplineInterpolator — Type

CubicSplineInterpolator(x::AbstractVector{<:Real}, y::AbstractVector{<:Real})

Construct a CubicSplineInterpolator for the points in x and y. This is a natural spline, where the second derivative is set to zero at the boundaries.

CubicSplineInterpolator(f::Function, xa::Real, xb::Real, n::Int)

Construct a CubicSplineInterpolator for the function f using n evenly spaced function evaluations in the range [xa,xb]

BasicInterpolators.BicubicSplineInterpolator — Type

BicubicSplineInterpolator(x::AbstractVector{<:Real},
                          y::AbstractVector{<:Real},
                          Z::AbstractArray{<:Real,2})

Construct a BicubicSplineInterpolator for the grid of points with axis 1 coordinates in x, axis 2 coordinates in y, and grid values in Z.

BicubicSplineInterpolator(f::Function,
                          xa::Real, xb::Real, nx::Int,
                          ya::Real, yb::Real, ny::Int)

Construct a BicubicSplineInterpolator for the function f using a grid of nx points evenly spaced on the first axis in [xa,xb] and ny points evenly spaced on the second axis in [ya,yb].