The following types can be passed to interpolator constructors to modify how points are handled at/outside the boundaries of the interpolation nodes.

BasicInterpolators.StrictBoundaries — Type

StrictBoundaries()

Triggers errors whenever interpolation coordinates are outside of the boundaries.

BasicInterpolators.WeakBoundaries — Type

WeakBoundaries()

Allows small overshoots at boundaries, but not large ones. Errors are only triggered when the interpolation coordinate is outside of a boundary and not close to it. At the lower boundar, for example, an error would be triggered when (x < boundary) & !(x ≈ boundary).

BasicInterpolators.NoBoundaries — Type

NoBoundaries()

Performs no boundary checking. Allows interpolators to blindly extrapolate when possible.